Twitch update

+++ lvlEdit/src/main/java/skullMod/lvlEdit/dataStructures/LVL/CameraTilt.java
@@ -6,11 +6,11 @@ public class CameraTilt implements Serializable{
 
 
     public float tiltRate;
-    public int tiltHeight1;
-    public int tiltHeight2;
+    public float tiltHeight1;
+    public float tiltHeight2;
 
 
-    public CameraTilt(float tiltRate, int tiltHeight1, int tiltHeight2){
+    public CameraTilt(float tiltRate, float tiltHeight1, float tiltHeight2){
         this.tiltRate = tiltRate;
         this.tiltHeight1 = tiltHeight1;
         this.tiltHeight2 = tiltHeight2;
